Move classes into guild package
