+++ src/main/java/net/cflip/bineclaims/BineClaims.java
@@ -1,6 +1,7 @@
 package net.cflip.bineclaims;
 
 import net.cflip.bineclaims.command.BineClaimsCommand;
+import net.cflip.bineclaims.guild.GuildManager;
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.command.v1.CommandRegistrationCallback;
 import net.fabricmc.fabric.api.event.lifecycle.v1.ServerLifecycleEvents;
