+++ src/main/java/net/cflip/bineclaims/command/BineClaimsCommand.java
@@ -5,7 +5,7 @@ import com.mojang.brigadier.StringReader;
 import com.mojang.brigadier.context.CommandContext;
 import com.mojang.brigadier.exceptions.CommandSyntaxException;
 import net.cflip.bineclaims.BineClaims;
-import net.cflip.bineclaims.Guild;
+import net.cflip.bineclaims.guild.Guild;
 import net.minecraft.server.command.CommandManager;
 import net.minecraft.server.command.ServerCommandSource;
